+++ src/freeze.ts
@@ -1,11 +1,14 @@
+import { DeepReadonly } from './utilities';
+
 /**
  * Perform shallow(_deep_ is `false`) or deep(_deep_ is `true`)
  * freeze of object. By default function does shallow freezing.
  */
-export function freeze<T extends object>(
+export function freeze<T extends object, D extends boolean = false>(
   value: T,
-  deep: boolean = false
-): Readonly<T> {
+  // @ts-ignore - TS cannot assing false to boolean :(
+  deep: D = false
+): D extends true ? DeepReadonly<T> : Readonly<T> {
   if (deep) {
     Object.getOwnPropertyNames(value).forEach((name) => {
       const innerValue = (value as { [key: string]: any })[name];
@@ -15,5 +18,6 @@ export function freeze<T extends object>(
     });
   }
 
+  // @ts-ignore
   return Object.freeze(value);
 }